Add files via upload
Creating an ensemble of classifiers to outperform individual classifiers. Minimal improvements in terms of accuracy. However, results are not reproducible and could be better.
To get similar results for your model on any run, use random state by scikit-learn or set the random seed of the cell you intend to run to a value of your choice.